refactoring (yes again)
