+++ server/controllers/moleculesImporter.js
@@ -5,9 +5,9 @@ import { HeaderErrors } from "../global/csv_reader/HeaderChecker.js";
 import { createDir, deleteFiles, getSortedFiles, moveFile } from "../global/Files.js";
 import HttpResponseWrapper from "../global/HttpResponseWrapper.js";
 import Logger, { addErrorTitle } from "../global/Logger.js";
-import { analyzeData } from "../global/molecules_analyzer/moleculesAnalyzer.js";
-import { createSqlToInsertAllData } from "../global/molecules_importer/moleculesImporter.js";
-import { parseMoleculesFromCsv } from "../global/molecules_parser/Parser.js";
+import { analyzeData } from "../global/molecules_importation/moleculesAnalyzer.js";
+import { createSqlToInsertAllData } from "../global/molecules_importation/moleculesImporter.js";
+import { parseMoleculesFromCsv } from "../global/molecules_importation/moleculesParser.js";
 
 const FILES_DIR_PATH = path.resolve("files", "molecules");
 const MAX_FILE_KEPT = 15;
