+++ server/test/molecules_analyzer/moleculesAnalyzer.test.js
@@ -6,9 +6,9 @@ import { queryPromise } from "../../db/database.js";
 import {
   analyzeData,
   MoleculesAnalyzerWarning,
-} from "../../global/molecules_analyzer/moleculesAnalyzer.js";
-import { createSqlToInsertAllData } from "../../global/molecules_importer/moleculesImporter.js";
-import { parseMoleculesFromCsv } from "../../global/molecules_parser/Parser.js";
+} from "../../global/molecules_importation/moleculesAnalyzer.js";
+import { createSqlToInsertAllData } from "../../global/molecules_importation/moleculesImporter.js";
+import { parseMoleculesFromCsv } from "../../global/molecules_importation/moleculesParser.js";
 import { forceTruncateTables } from "../index.test.js";
 
 const { expect } = chai;
