+++ server/test/molecules_parser/moleculesParser.test.js
@@ -7,8 +7,8 @@ import mocha from "mocha";
 
 import { HeaderErrors } from "../../global/csv_reader/HeaderChecker.js";
 // eslint-disable-next-line no-unused-vars
-import { ClassificationNode } from "../../global/molecules_parser/MoleculesClassification.js";
-import { parseMoleculesFromCsv } from "../../global/molecules_parser/Parser.js";
+import { ClassificationNode } from "../../global/molecules_importation/MoleculesClassification.js";
+import { parseMoleculesFromCsv } from "../../global/molecules_importation/moleculesParser.js";
 
 import { expectations } from "./expectations.js";
